Refactor Intake.ChangeIntakePosition to use switch
Since there are only 2 possible options, an if-else if-else chain is unnecessary.
This refactors it to use switch/case
Using switch/case instead of if/else automatically makes it so that whenever a new position is added to the enum,
the compile will error unless it is also added to the switch/case.